+++ R/utils.R
@@ -6,7 +6,7 @@
 #' @return character string
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #'  doc_path(filename = "my_tbl.pdf", path = NULL)
 #'}
 doc_path <- function(filename = NULL, path = NULL){
@@ -180,4 +180,29 @@ create_chunk <- function(x, transform) {
   ")
 
   cat(knitr::knit(text = knitr::knit_expand(text = new_chunk), quiet = TRUE))
+}
+
+
+#' Check package versions for docorator object are the same as loaded
+#'
+#' @param x docorator object
+#' @keywords internal
+check_pkg_version <- function(x) {
+
+  if (rlang::inherits_any(x$display, c("gt_tbl", "gt_group"))) {
+    pkg <- "gt"
+  }else if (inherits(x$display,"ggplot")) {
+    pkg <- "ggplot2"
+  }else{
+    return()
+  }
+
+ package_list <- append(x$session_info$otherPkgs,x$session_info$loadedOnly)
+ pkg_version <- package_list[[pkg]][["Version"]]
+
+ # get current version if it exists
+ current_version = as.character(utils::packageVersion(pkg))
+
+ if(utils::compareVersion(current_version, pkg_version)!= 0){
+   cli::cli_text("Note: docorator object was created with {.pkg {pkg} {pkg_version}}. You are now running {.pkg {pkg} {current_version}}. There may be issues rendering your document.")}
 }